third_party/WebKit/LayoutTests/shadow-dom/slots-fallback-api-4.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/shadow-dom.js"></script>
<div id="d1">
  <template data-mode="open">
    <div id="d1_d1">
      <template data-mode="open">
        <slot id="d1_d1_s1" name="d1_d1_s1"></slot>
      </template>
      <slot id="d1_s1" name="d1_s1" slot="d1_d1_s1"></slot>
    </div>
  </template>
</div>
<script>
'use strict';

let n = createTestTree(d1);
removeWhiteSpaceOnlyTextNodes(n.d1);

test(() => {
  assert_array_equals(n.d1_s1.assignedNodes(), []);
  assert_array_equals(n.d1_s1.assignedNodes({ flatten: true }), []);
  assert_array_equals(n.d1_d1_s1.assignedNodes(), [n.d1_s1]);
  assert_array_equals(n.d1_d1_s1.assignedNodes({ flatten: true }), []);

  let d1_d2 = document.createElement('div');
  d1_d2.setAttribute('id', 'd1_d2');
  n.d1_s1.appendChild(d1_d2);

  assert_array_equals(n.d1_s1.assignedNodes({ flatten: true }), [d1_d2]);
  assert_array_equals(n.d1_d1_s1.assignedNodes({ flatten: true }), [d1_d2]);
}, "Distribution should be updated after mutating fallback contents");
</script>
